requirements.txt
pandas==2.1.1
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
mlflow==2.9.2
kafka-python==2.1.0
